requirements.txt
yfinance
fredapi==0.5.1
great-expectations==0.18.19
pandas==2.2.2
numpy==1.26.4
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
